pkginfo: footprint mode now makes only one pass through the archive
TL;DR:
1. This cuts the average amount of time required in half.
2. This requires to remake all packages' footprint files.

Overview:

James Buren <ryuo at ryuo.xyz> improved footprint mode in pkginfo[1].
The original approach wastes time by making two passes through each
archive when generating a footprint.  The revised algorithm caches the
necessary metadata it needs to generate the footprint in a first pass
through the archive and then process it all during a second pass
through the cached metadata.

The observable differences from the original algorithm are as follows.
First, the new algorithm takes roughly half the time of the original
when generating footprints.  Second, the unsorted line order has been
changed from the order of its appearance in the original archive to a
sorted order by path.  The last change was done to support the faster
binary search lookups for the hard links that the original was doing
via a hash table.

If this is considered undesirable, it can be switched back to using a
hash table to retain the original order.  It was switched to not
increase the RAM usage anymore that already was for the metadata cache.
An auxiliary hashtable would require extra RAM whereas this method
allows to reuse the same array that was used for metadata cache.

AS A CONSEQUENCES, WITH PKGUTILS UPDATE, ALL FOOTPRINT FILES NEED TO
BE REGENERATED.  This can be done as follows:

  pkginfo -l PACKAGE > .footprint

BE CAREFUL WHEN UPDATING!

References:
[1]: https://lists.crux.nu/pipermail/crux-devel/2019-June/004107.html